trying to fix issue with session.json failing to load

+++ withings_sync/python_garminconnect/example.py
@@ -496,7 +496,8 @@ def uploadWithingsData(garmin_username, garmin_password, email_server, email_ema
         api.upload_activity(activityfile)
     except:
         print("Failed to use Session. Trying login with MFA")
-        os.remove("session.json")
+        if os.path.exists("session.json"):
+            os.remove("session.json")
         api = init_api(garmin_username, garmin_password)
         api.upload_activity(activityfile)
